+++ general/communication.hpp
@@ -431,7 +431,7 @@ struct VarMessage
    void Isend(int rank, MPI_Comm comm)
    {
       Encode(rank);
-      MPI_Isend((void*) data.data(), data.length(), MPI_BYTE, rank, Tag, comm,
+      MPI_Isend((void*) data.data(), static_cast<int>(data.length()), MPI_BYTE, rank, Tag, comm,
                 &send_request);
    }
 
@@ -441,7 +441,7 @@ struct VarMessage
    void Issend(int rank, MPI_Comm comm)
    {
       Encode(rank);
-      MPI_Issend((void*) data.data(), data.length(), MPI_BYTE, rank, Tag, comm,
+      MPI_Issend((void*) data.data(), static_cast<int>(data.length()), MPI_BYTE, rank, Tag, comm,
                  &send_request);
    }
 
@@ -538,7 +538,7 @@ struct VarMessage
    template<typename MapT>
    static void RecvAll(MapT& rank_msg, MPI_Comm comm)
    {
-      int recv_left = rank_msg.size();
+      int recv_left = static_cast<int>(rank_msg.size());
       while (recv_left > 0)
       {
          int rank, size;
